Move the fake log to a better folder
