+++ test/fake/fakeLog.ts
