+++ test/plugin/plugin.test.ts
@@ -10,7 +10,7 @@ import fetch from 'node-fetch'
 
 import edgeCorePlugins from '../../src/index'
 import { expectRejection } from '../expectRejection'
-import { fakeLog } from '../fakeLog'
+import { fakeLog } from '../fake/fakeLog'
 import fixtures from './fixtures'
 
 for (const fixture of fixtures) {
